+++ src/Phan/LanguageServer/CompletionRequest.php
@@ -63,8 +63,11 @@ final class CompletionRequest extends NodeInfoRequest
     public function recordCompletionList($completions)
     {
         if ($completions instanceof CompletionItem || isset($completions['label'])) {
+            // Written this way so that Phan can infer $completions is not a CompletionItem
+            // FIXME: There was a regression in the analysis of negation of `instanceof`
             $completions = [$completions];
         }
+        // @phan-suppress-next-line PhanTypeSuspiciousNonTraversableForeach
         foreach ($completions ?? [] as $completion) {
             if (is_array($completion)) {
                 $completion = CompletionItem::fromArray($completion);
@@ -176,39 +179,39 @@ final class CompletionRequest extends NodeInfoRequest
 
     public function finalize()
     {
-        $promise = $this->promise;
-        if ($promise) {
-            $result = $this->completions ?: null;
-            if ($result !== null) {
-                // Sort completion suggestions alphabetically,
-                // ignoring the leading `$` in variables/static properties.
-                \uksort(
-                    $result,
-                    /**
-                     * @param string $a
-                     * @param string $b
-                     */
-                    static function ($a, $b) : int {
-                        $a = \ltrim((string)$a, '$');
-                        $b = \ltrim((string)$b, '$');
-                        return (\strtolower($a) <=> \strtolower($b)) ?: ($a <=> $b);
-                    }
-                );
-                $result_list = new CompletionList(\array_values($result));
-            } else {
-                $result_list = null;
-            }
-            $promise->fulfill($result_list);
-            $this->promise = null;
+        if ($this->fulfilled) {
+            return;
         }
+        $this->fulfilled = true;
+        $result = $this->completions ?: null;
+        if ($result !== null) {
+            // Sort completion suggestions alphabetically,
+            // ignoring the leading `$` in variables/static properties.
+            \uksort(
+                $result,
+                /**
+                 * @param string $a
+                 * @param string $b
+                 */
+                static function ($a, $b) : int {
+                    $a = \ltrim((string)$a, '$');
+                    $b = \ltrim((string)$b, '$');
+                    return (\strtolower($a) <=> \strtolower($b)) ?: ($a <=> $b);
+                }
+            );
+            $result_list = new CompletionList(\array_values($result));
+        } else {
+            $result_list = null;
+        }
+        $this->promise->fulfill($result_list);
     }
 
     public function __destruct()
     {
-        $promise = $this->promise;
-        if ($promise) {
-            $promise->reject(new Exception('Failed to send a valid textDocument/definition result'));
-            $this->promise = null;
+        if ($this->fulfilled) {
+            return;
         }
+        $this->fulfilled = true;
+        $this->promise->reject(new Exception('Failed to send a valid textDocument/definition result'));
     }
 }